requirements.txt
fastapi>=0.110,<1
uvicorn>=0.29,<1
SQLAlchemy>=2,<3
python-jose[cryptography]>=3,<4
passlib[bcrypt]>=1.7,<2
python-multipart
python-dotenv>=1,<2
pandas>=2,<3
numpy>=1.26,<3
openpyxl>=3.1,<4
